internal: refactor unresolved extern crate diagnostic
